Converted script loader to custom hook
